requirements.txt
django==4.1.3            
django-rest-knox==4.2.0
django-rest-framework==3.14.0